Directional lighting

+++ src/slap42/graphics/meshes/mesh.hpp
@@ -14,7 +14,7 @@ public:
 
 private:
   unsigned int vao, vbo, ibo;
-  int index_count;
+  size_t index_count;
 };
 
 }
